backend/requirements.txt
alembic==1.3.1
astroid==2.2.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.11
twilio==6.33.1
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
